+++ android-app/app/src/main/java/com/echopen/asso/echopen/ChoiceActivity.java
@@ -1,9 +1,10 @@
 package com.echopen.asso.echopen;
 
+import android.app.Activity;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class ChoiceActivity extends AppCompatActivity {
+public class ChoiceActivity extends Activity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
