+++ android-app/app/src/main/res/layout/activity_main.xml
@@ -12,11 +12,29 @@
         android:text="@string/app_name"
         />
 
-    <ImageView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:id = '@+id/img'
-        />
+    <Button
+        android:id="@+id/enter_bt"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Enter"/>
+
+    <Button
+        android:id="@+id/signin_bt"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="sign in"/>
+
+    <Button
+        android:id="@+id/signout_bt"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="sign out"/>
+
+    <Button
+        android:id="@+id/emergency_bt"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Emergency"/>
 
 </LinearLayout>
